+++ src/pyramid_bimt/tests/test_security.py
@@ -26,3 +26,33 @@ class TestVerify(unittest.TestCase):
         self.assertEqual(len(handler.records), 1)
         self.assertEqual(
             handler.records[0].message, 'hash could not be identified')
+
+    def test_verify(self):
+        from pyramid_bimt.security import generate
+        from pyramid_bimt.security import encrypt
+        from pyramid_bimt.security import verify
+
+        generated_pass = generate()
+        encrypted_pass = encrypt(generated_pass)
+
+        self.assertTrue(verify(generated_pass, encrypted_pass))
+
+    def test_verify_none(self):
+        from pyramid_bimt.security import generate
+        from pyramid_bimt.security import encrypt
+        from pyramid_bimt.security import verify
+
+        generated_pass = generate()
+        encrypted_pass = encrypt(generated_pass)
+
+        self.assertFalse(verify(None, encrypted_pass))
+
+    def test_verify_wrong_type(self):
+        from pyramid_bimt.security import generate
+        from pyramid_bimt.security import encrypt
+        from pyramid_bimt.security import verify
+
+        generated_pass = generate()
+        encrypted_pass = encrypt(generated_pass)
+
+        self.assertFalse(verify(object(), encrypted_pass))